Organized header files for the libhyperdbg project
