+++ hyperdbg/libhyperdbg/header/app/messaging.h
