+++ hyperdbg/libhyperdbg/header/common/common.h
