+++ hyperdbg/libhyperdbg/header/common/list.h
