+++ hyperdbg/libhyperdbg/header/debugger/commands/commands.h
