+++ hyperdbg/libhyperdbg/header/debugger/commands/help.h
