+++ hyperdbg/libhyperdbg/header/debugger/communication/communication.h
