+++ hyperdbg/libhyperdbg/header/debugger/communication/forwarding.h
