+++ hyperdbg/libhyperdbg/header/debugger/communication/namedpipe.h
