+++ hyperdbg/libhyperdbg/header/debugger/core/debugger.h
