+++ hyperdbg/libhyperdbg/header/debugger/kernel-level/kd.h
