+++ hyperdbg/libhyperdbg/header/debugger/script-engine/script-engine.h
