+++ hyperdbg/libhyperdbg/header/debugger/script-engine/symbol.h
