+++ hyperdbg/libhyperdbg/header/debugger/transparency/transparency.h
