+++ hyperdbg/libhyperdbg/header/debugger/user-level/pe-parser.h
