+++ hyperdbg/libhyperdbg/header/globals/globals.h
