+++ hyperdbg/libhyperdbg/header/hwdbg/hwdbg-interpreter.h
