+++ hyperdbg/libhyperdbg/header/hwdbg/hwdbg-scripts.h
