+++ hyperdbg/libhyperdbg/header/objects/objects.h
