+++ hyperdbg/libhyperdbg/libhyperdbg.vcxproj
@@ -140,35 +140,35 @@ msbuild "$(SolutionDir)dependencies\zydis\msvc\Zydis.sln" /m /p:Configuration="R
     <ClInclude Include="..\include\platform\general\header\nt-list.h" />
     <ClInclude Include="..\include\platform\user\header\windows-only\windows-privilege.h" />
     <ClInclude Include="..\include\platform\user\header\Windows.h" />
-    <ClInclude Include="header\assembler.h" />
-    <ClInclude Include="header\commands.h" />
-    <ClInclude Include="header\common.h" />
-    <ClInclude Include="header\communication.h" />
-    <ClInclude Include="header\debugger.h" />
-    <ClInclude Include="header\export.h" />
-    <ClInclude Include="header\forwarding.h" />
-    <ClInclude Include="header\globals.h" />
-    <ClInclude Include="header\help.h" />
-    <ClInclude Include="header\hwdbg-interpreter.h" />
-    <ClInclude Include="header\hwdbg-scripts.h" />
-    <ClInclude Include="header\inipp.h" />
-    <ClInclude Include="header\install.h" />
-    <ClInclude Include="header\kd.h" />
-    <ClInclude Include="header\libhyperdbg.h" />
-    <ClInclude Include="header\list.h" />
-    <ClInclude Include="header\messaging.h" />
-    <ClInclude Include="header\namedpipe.h" />
-    <ClInclude Include="header\objects.h" />
-    <ClInclude Include="header\packets.h" />
-    <ClInclude Include="header\pci-id.h" />
-    <ClInclude Include="header\pe-parser.h" />
-    <ClInclude Include="header\rev-ctrl.h" />
-    <ClInclude Include="header\script-engine.h" />
-    <ClInclude Include="header\steppings.h" />
-    <ClInclude Include="header\symbol.h" />
-    <ClInclude Include="header\tests.h" />
-    <ClInclude Include="header\transparency.h" />
-    <ClInclude Include="header\ud.h" />
+    <ClInclude Include="header\app\libhyperdbg.h" />
+    <ClInclude Include="header\app\messaging.h" />
+    <ClInclude Include="header\app\packets.h" />
+    <ClInclude Include="header\common\common.h" />
+    <ClInclude Include="header\common\list.h" />
+    <ClInclude Include="header\debugger\commands\commands.h" />
+    <ClInclude Include="header\debugger\commands\help.h" />
+    <ClInclude Include="header\debugger\communication\communication.h" />
+    <ClInclude Include="header\debugger\communication\forwarding.h" />
+    <ClInclude Include="header\debugger\communication\namedpipe.h" />
+    <ClInclude Include="header\debugger\core\debugger.h" />
+    <ClInclude Include="header\debugger\core\steppings.h" />
+    <ClInclude Include="header\debugger\driver-loader\install.h" />
+    <ClInclude Include="header\debugger\kernel-level\kd.h" />
+    <ClInclude Include="header\debugger\misc\assembler.h" />
+    <ClInclude Include="header\debugger\misc\inipp.h" />
+    <ClInclude Include="header\debugger\misc\pci-id.h" />
+    <ClInclude Include="header\debugger\script-engine\script-engine.h" />
+    <ClInclude Include="header\debugger\script-engine\symbol.h" />
+    <ClInclude Include="header\debugger\tests\tests.h" />
+    <ClInclude Include="header\debugger\transparency\transparency.h" />
+    <ClInclude Include="header\debugger\user-level\pe-parser.h" />
+    <ClInclude Include="header\debugger\user-level\ud.h" />
+    <ClInclude Include="header\export\export.h" />
+    <ClInclude Include="header\globals\globals.h" />
+    <ClInclude Include="header\hwdbg\hwdbg-interpreter.h" />
+    <ClInclude Include="header\hwdbg\hwdbg-scripts.h" />
+    <ClInclude Include="header\objects\objects.h" />
+    <ClInclude Include="header\rev\rev-ctrl.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
