+++ hyperdbg/libhyperdbg/pch.h
@@ -187,7 +187,7 @@ typedef const wchar_t *LPCWCHAR, *PCWCHAR;
 //
 // PCI IDs
 //
-#include "header/pci-id.h"
+#include "header/debugger/misc/pci-id.h"
 
 //
 // Intel PT
@@ -197,45 +197,45 @@ typedef const wchar_t *LPCWCHAR, *PCWCHAR;
 //
 // General
 //
-#include "header/libhyperdbg.h"
-#include "header/export.h"
-#include "header/inipp.h"
-#include "header/commands.h"
-#include "header/common.h"
-#include "header/symbol.h"
-#include "header/debugger.h"
-#include "header/script-engine.h"
-#include "header/help.h"
+#include "header/app/libhyperdbg.h"
+#include "header/export/export.h"
+#include "header/debugger/misc/inipp.h"
+#include "header/debugger/commands/commands.h"
+#include "header/common/common.h"
+#include "header/debugger/script-engine/symbol.h"
+#include "header/debugger/core/debugger.h"
+#include "header/debugger/script-engine/script-engine.h"
+#include "header/debugger/commands/help.h"
 #ifdef _WIN32
-#    include "header/install.h"
+#    include "header/debugger/driver-loader/install.h"
 #endif
-#include "header/list.h"
-#include "header/tests.h"
-#include "header/messaging.h"
-#include "header/packets.h"
-#include "header/transparency.h"
-#include "header/communication.h"
-#include "header/namedpipe.h"
-#include "header/forwarding.h"
-#include "header/kd.h"
+#include "header/common/list.h"
+#include "header/debugger/tests/tests.h"
+#include "header/app/messaging.h"
+#include "header/app/packets.h"
+#include "header/debugger/transparency/transparency.h"
+#include "header/debugger/communication/communication.h"
+#include "header/debugger/communication/namedpipe.h"
+#include "header/debugger/communication/forwarding.h"
+#include "header/debugger/kernel-level/kd.h"
 
 //
 // Components
 //
 #include "../include/components/pe/header/pe-image-reader.h"
 
-#include "header/pe-parser.h"
-#include "header/ud.h"
-#include "header/objects.h"
-#include "header/steppings.h"
-#include "header/rev-ctrl.h"
-#include "header/assembler.h"
+#include "header/debugger/user-level/pe-parser.h"
+#include "header/debugger/user-level/ud.h"
+#include "header/objects/objects.h"
+#include "header/debugger/core/steppings.h"
+#include "header/rev/rev-ctrl.h"
+#include "header/debugger/misc/assembler.h"
 
 //
 // hwdbg
 //
-#include "header/hwdbg-interpreter.h"
-#include "header/hwdbg-scripts.h"
+#include "header/hwdbg/hwdbg-interpreter.h"
+#include "header/hwdbg/hwdbg-scripts.h"
 
 //
 // Zydis headers
